+++ shop/settings.py
@@ -31,7 +31,6 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'core',
     'accounts',
-    'mall',
     'finance',
     'bot',
     'orders',